+++ bin/coreDataTest.sh
@@ -2,8 +2,7 @@
 
 NAMESFILE=$(dirname "$0")/files.sh
 
-COLLECTION_PATH="collections/core-data.postman_collection.json"
-ENV_PATH="environment/CoredataEnv.postman_environment.json"
+_TESTMODE=${_TESTMODE:docker}
 
 if [ -f $NAMESFILE ]; then 
 
@@ -17,31 +16,62 @@ fi
 
 echo "Info: Initiating Coredata Test."
 
-echo "[info] ---------- use docker-compose run newman ----------"
+if [ "$_TESTMODE" == "docker" ] ; then
 
-docker-compose run --rm postman run ${COLLECTION_PATH} \
-    --folder="event" --iteration-data="data/eventData.json" --environment=${ENV_PATH} \
-    --reporters="junit,cli"
-docker-compose run --rm postman run ${COLLECTION_PATH} \
-    --folder="event_error_4xx" --iteration-data="data/eventData.json" --environment=${ENV_PATH} \
-    --reporters="junit,cli"
+    COLLECTION_PATH="collections/core-data.postman_collection.json"
+    ENV_PATH="environment/CoredataEnv.postman_environment.json"
 
-docker-compose run --rm postman run ${COLLECTION_PATH} \
-    --folder="reading" --iteration-data="data/readingData.json" --environment=${ENV_PATH} \
-    --reporters="junit,cli"
-docker-compose run --rm postman run ${COLLECTION_PATH} \
-    --folder="reading_error_4xx" --iteration-data="data/readingData.json" --environment=${ENV_PATH} \
-    --reporters="junit,cli"
+    echo "[info] ---------- use docker-compose run postman ----------"
 
-docker-compose run --rm postman run ${COLLECTION_PATH} \
-    --folder="valuedescriptor" --iteration-data="data/valueDescriptorData.json" --environment=${ENV_PATH} \
-    --reporters="junit,cli"
-docker-compose run --rm postman run ${COLLECTION_PATH} \
-    --folder="valuedescriptor_error_4xx" --iteration-data="data/valueDescriptorData.json" --environment=${ENV_PATH} \
-    --reporters="junit,cli"
+    docker-compose run --rm postman run ${COLLECTION_PATH} \
+        --folder="event" --iteration-data="data/eventData.json" --environment=${ENV_PATH} \
+        --reporters="junit,cli"
+    docker-compose run --rm postman run ${COLLECTION_PATH} \
+        --folder="event_error_4xx" --iteration-data="data/eventData.json" --environment=${ENV_PATH} \
+        --reporters="junit,cli"
 
+    docker-compose run --rm postman run ${COLLECTION_PATH} \
+        --folder="reading" --iteration-data="data/readingData.json" --environment=${ENV_PATH} \
+        --reporters="junit,cli"
+    docker-compose run --rm postman run ${COLLECTION_PATH} \
+        --folder="reading_error_4xx" --iteration-data="data/readingData.json" --environment=${ENV_PATH} \
+        --reporters="junit,cli"
 
+    docker-compose run --rm postman run ${COLLECTION_PATH} \
+        --folder="valuedescriptor" --iteration-data="data/valueDescriptorData.json" --environment=${ENV_PATH} \
+        --reporters="junit,cli"
 
+    docker-compose run --rm postman run ${COLLECTION_PATH} \
+        --folder="valuedescriptor_error_4xx" --iteration-data="data/valueDescriptorData.json" --environment=${ENV_PATH} \
+        --reporters="junit,cli"
+
+elif [ "$_TESTMODE" == "newman" ] ; then
+
+    COLLECTION_PATH="./postman-test/collections/core-data.postman_collection.json"
+    ENV_PATH="./postman-test/environment/CoredataEnv.postman_environment.json"
+
+    echo "[info] ---------- use newman ----------"
+
+    newman run ${COLLECTION_PATH} \
+	   --folder event -e ${ENV_PATH} \
+	   -d ./postman-test/data/eventData.json \
+	   -r cli
+
+    newman run ${COLLECTION_PATH} \
+	   --folder event_error_4xx -e ${ENV_PATH} \
+	   -d ./postman-test/data/eventData.json \
+	   -r cli
+
+    newman run ${COLLECTION_PATH} \
+	   --folder reading -e ${ENV_PATH} \
+	   -d ./postman-test/data/readingData.json \
+	   -r cli
+
+    newman run ${COLLECTION_PATH} \
+	   --folder reading_error_4xx -e ${ENV_PATH} \
+	   -d ./postman-test/data/eventData.json \
+	   -r cli
+fi
 
 #docker run --rm --user="1000" -v "${PWD}/bin/postman-test":/etc/newman --network=${DOCKER_NETWORK} postman/newman_ubuntu1404 run ${COLLECTION_PATH} \
 #    --folder="event" --iteration-data="data/eventData.json" --environment=${ENV_PATH} \
